+++ argopy/data_fetchers/argovis_data.py
@@ -15,12 +15,11 @@ import getpass
 from .proto import ArgoDataFetcherProto
 from abc import abstractmethod
 import warnings
-import urllib
 
 from argopy.stores import httpstore
 from argopy.options import OPTIONS
 from argopy.utilities import list_standard_variables, format_oneline, Chunker
-from argopy.plotters import open_dashboard
+
 
 access_points = ['wmo', 'box']
 exit_formats = ['xarray']
@@ -433,12 +432,6 @@ class Fetch_wmo(ArgovisDataFetcher):
 
         urls = list_bunch(self.WMO, self.CYC)
         return self.url_encode(urls)
-
-    def dashboard(self, **kw):
-        if len(self.WMO) == 1:
-            return open_dashboard(wmo=self.WMO[0], **kw)
-        else:
-            warnings.warn("Plot dashboard only available for request with a single float")
 
 
 class Fetch_box(ArgovisDataFetcher):