+++ argopy/data_fetchers/erddap_data.py
@@ -13,7 +13,6 @@ This is not intended to be used directly, only by the facade at fetchers.py
 import pandas as pd
 import numpy as np
 import copy
-import warnings
 
 from abc import abstractmethod
 import getpass
@@ -22,7 +21,6 @@ from .proto import ArgoDataFetcherProto
 from argopy.options import OPTIONS
 from argopy.utilities import list_standard_variables, Chunker, format_oneline
 from argopy.stores import httpstore
-from argopy.plotters import open_dashboard
 
 
 # Load erddapy according to available version (breaking changes in v0.8.0)
@@ -599,12 +597,6 @@ class Fetch_wmo(ErddapArgoDataFetcher):
                 ).get_url()
             )
         return urls
-
-    def dashboard(self, **kw):
-        if len(self.WMO) == 1:
-            return open_dashboard(wmo=self.WMO[0], **kw)
-        else:
-            warnings.warn("Dashboard only available for a single float request")
 
 
 class Fetch_box(ErddapArgoDataFetcher):