+++ argopy/tests/test_fetchers_facade_data.py
@@ -102,6 +102,7 @@ class Test_Facade:
         with argopy.set_options(local_ftp=self.local_ftp):
             assert isinstance(self.__get_fetcher()[1].to_index(), pd.core.frame.DataFrame)
             assert isinstance(self.__get_fetcher()[1].to_index(full=True), pd.core.frame.DataFrame)
+            assert isinstance(self.__get_fetcher()[1].to_index(full=False, coriolis_id=True), pd.core.frame.DataFrame)
 
     def test_load(self):
         with argopy.set_options(local_ftp=self.local_ftp):
@@ -167,6 +168,7 @@ class Test_Facade:
             # Test 'qc_altimetry'
             dsh = fetcher.plot(ptype='qc_altimetry', embed='slide')
             assert isinstance(dsh(0), IPython.display.Image)
+
 
 @requires_fetcher
 class Test_DataFetching: