+++ argopy/tests/test_fetchers_facade_index.py
@@ -1,5 +1,6 @@
 import pandas as pd
 import pytest
+import importlib
 
 import argopy
 from argopy import IndexFetcher as ArgoIndexFetcher
@@ -125,8 +126,6 @@ class Test_Facade:
         import IPython
         with argopy.set_options(local_ftp=self.local_ftp):
             f, fetcher = self.__get_fetcher(pt='float')
-
-            # Test 'qc_altimetry'
             dsh = fetcher.plot(ptype='qc_altimetry', embed='slide')
             assert isinstance(dsh(0), IPython.display.Image)