+++ docs/whats-new.rst
@@ -23,6 +23,35 @@ v0.1.11 (X XXX. 2022)
 
 **Features and front-end API**
 
+- **New dashboard for profiles and new 3rd party dashboards**. Calling on the data fetcher dashboard method will return the Euro-Argo profile page for a single profile. Very useful to look at the data before load. This comes with 2 new utilities functions to get Coriolis ID of profiles (:meth:`utilities.get_coriolis_profile_id`) and to return the list of profile webpages (:meth:`utilities.get_ea_profile_page`). (:pr:`198`) by `G. Maze <http://www.github.com/gmaze>`_.
+
+.. code-block:: python
+
+    from argopy import DataFetcher as ArgoDataFetcher
+    ArgoDataFetcher().profile(5904797, 11).dashboard()
+
+.. code-block:: python
+
+    from argopy.utilities import get_coriolis_profile_id, get_ea_profile_page
+    get_coriolis_profile_id([6902755, 6902756], [11, 12])
+    get_ea_profile_page([6902755, 6902756], [11, 12])
+
+The new profile dashboard can also be accessed with:
+
+.. code-block:: python
+
+    import argopy
+    argopy.dashboard(5904797, 11)
+
+We added the Ocean-OPS (former JCOMMOPS) dashboard for all floats and the Argo-BGC dashboard for BGC floats:
+
+.. code-block:: python
+
+    import argopy
+    argopy.dashboard(5904797, type='ocean-ops')
+    # or
+    argopy.dashboard(5904797, 12, type='bgc')
+
 - New utility function :class:`argopy.utilities.ArgoNVSReferenceTables` to retrieve Argo Reference Tables. (:commit:`cc8fdbe132874b71b35203053626cc29ae7d19c4`) by `G. Maze <http://www.github.com/gmaze>`_.
 
 .. code-block:: python
@@ -36,6 +65,13 @@ v0.1.11 (X XXX. 2022)
 **Internals**
 
 - Fix bug in erddap fata fetcher that was causing a `profile` request to do not account for cycle numbers. (:commit:`301e557fdec1f2d536841464b383edc3a4c4a62d`) by `G. Maze <http://www.github.com/gmaze>`_.
+
+- Complete refactoring of the ``argopy.plotters`` module into ``argopy.plot``. (:pr:`198`) by `G. Maze <http://www.github.com/gmaze>`_.
+
+**Breaking changes with previous versions**
+
+- Remove deprecation warnings for: 'plotters.plot_dac', 'plotters.plot_profilerType'. These now raise an error.
+- ``argopy.plotters`` module replaced by ``argopy.plot``
 
 
 v0.1.10 (4 Mar. 2022)